requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid==0.7.0
asteroid-filterbanks==0.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audiomentations==0.24.0
audioread==3.0.1
auraloss==0.4.0
av==14.3.0
backports-datetime-fromisoformat==2.0.3
beartype==0.14.1
bitsandbytes==0.45.5
braceexpand==0.1.7
cached-property==2.0.1
cachetools==5.5.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
ctc-forced-aligner @ git+https://github.com/MahmoudAshraf97/ctc-forced-aligner.git@201276a4ea2ddd3f5caead1ac4f211477ae3da6d
ctranslate2==4.4.0
cycler==0.12.1
Cython==0.29.37
cytoolz==1.0.1
datasets==3.2.0
decorator==5.2.1
deepmultilingualpunctuation @ git+https://github.com/oliverguhr/deepmultilingualpunctuation.git@5a0dd7f4fd56687f59405aa8eba1144393d8b74b
demucs==4.0.0
diffq==0.2.4
dill==0.3.8
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
dora_search==0.1.12
editdistance==0.8.1
efficientnet_pytorch==0.7.1
einops==0.8.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.2.0
faster-whisper==1.1.1
ffmpeg-python==0.2.0
fiddle==0.3.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.9.0
future==1.0.0
g2p-en==2.1.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-genai==1.13.0
graphviz==0.20.3
greenlet==3.2.1
grpcio==1.71.0
h11==0.16.0
hdbscan==0.8.37
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
hydra-core==1.3.2
hyper-connections==0.1.11
HyperPyYAML==1.2.2
idna==3.10
inflect==7.5.0
intervaltree==3.1.0
ipython==8.36.0
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.4.2
julius==0.2.7
jupyterlab_widgets==3.0.14
kaldi-python-io==1.2.2
kaldiio==2.18.1
keyboard==0.13.5
kiwisolver==1.4.8
lameenc==1.8.1
lazy_loader==0.4
Levenshtein==0.27.1
lhotse==1.31.0
libcst==1.7.0
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.14.3
lilcom==1.8.1
llvmlite==0.44.0
loguru==0.7.3
loralib==0.1.2
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.1.6
mir_eval==0.8.2
ml_collections==1.1.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
munch==4.0.0
nemo-toolkit==2.2.1
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.2.3
onnx==1.17.0
onnxruntime==1.21.1
openpyxl==3.1.2
openunmix==1.3.0
optuna==4.3.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pb-bss-eval==0.0.2
pedalboard==0.8.9
peft==0.15.2
pesq==0.0.4
pexpect==4.9.0
pillow==11.2.1
plac==1.4.5
platformdirs==4.3.7
pooch==1.8.2
pretrainedmodels==0.7.4
primePy==1.3
prodigyopt==1.1.2
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pyloudnorm==0.1.1
pynndescent==0.5.13
pyparsing==3.2.3
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.5.1.post0
pytorch-ranger==0.1.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
retrying==1.3.4
rich==14.0.0
rotary-embedding-torch==0.3.5
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sacremoses==0.1.1
safetensors==0.5.3
sageattention==1.0.6
scikit-learn==1.6.1
scipy==1.15.2
segmentation-models-pytorch==0.3.3
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.5
shellingham==1.5.4
silero-vad==5.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==0.5.0.post1
spafe==0.3.2
speechbrain==1.0.3
SQLAlchemy==2.0.40
stack-data==0.6.3
submitit==1.5.2
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
text-unidecode==1.3
texterrors==0.5.1
threadpoolctl==3.6.0
timm==0.9.2
tokenizers==0.21.1
toolz==1.0.0
torch==2.7.0
torch-audiomentations==0.12.0
torch-log-wmse==0.3.0
torch-optimizer==0.1.0
torch-stoi==0.2.3
torch_pitch_shift==1.2.5
torchaudio==2.7.0
torchmetrics==0.11.4
torchseg==0.0.1a1
torchvision==0.22.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
treetable==0.2.5
triton==3.3.0
typeguard==4.4.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.6
Unidecode==1.4.0
urllib3==2.4.0
wandb==0.19.10
wcwidth==0.2.13
webdataset==0.2.111
websockets==15.0.1
Werkzeug==3.1.3
wespeaker @ git+https://github.com/wenet-e2e/wespeaker.git@310a15850895b54e20845e107b54c9a275d39a2d
wget==3.2
widgetsnbextension==4.0.14
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
